requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai-whisper==20250625
pyttsx3==2.90
pydub==0.25.1
soundfile==0.12.1
numpy==1.24.3
librosa==0.10.0
python-multipart==0.0.6
